@@ -13,7 +13,8 @@ repository = "https://github.com/openlawlibrary/pygls"
 documentation = "https://pygls.readthedocs.io/en/latest"
 license = "Apache-2.0"
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
+
 dependencies = [
     "attrs >= 24.3.0",
     "cattrs >= 23.1.2",
