@@ -30,7 +30,6 @@ to when they are required.
 """
 
 import re
-from typing import Optional
 
 from lsprotocol import types
 from pygls.cli import start_server
@@ -40,7 +39,7 @@ NUMBER = re.compile(r"\d+")
 server = LanguageServer("inlay-hint-server", "v1")
 
 
-def parse_int(chars: str) -> Optional[int]:
+def parse_int(chars: str) -> int | None:
     try:
         return int(chars)
     except Exception:
