@@ -23,7 +23,6 @@ import pytest_asyncio
 from lsprotocol import types
 
 if typing.TYPE_CHECKING:
-    from typing import Tuple
 
     from pygls.lsp.client import BaseLanguageClient
 
@@ -45,7 +44,7 @@ async def async_shutdown(runtime: str, get_client_for):
 
 
 async def test_async_shutdown(
-    async_shutdown: Tuple[BaseLanguageClient, types.InitializeResult], uri_for
+    async_shutdown: tuple[BaseLanguageClient, types.InitializeResult], uri_for
 ):
     """Ensure that the formatting provider is not set by default."""
     client, _ = async_shutdown
