@@ -23,7 +23,6 @@ import pytest_asyncio
 from lsprotocol import types
 
 if typing.TYPE_CHECKING:
-    from typing import Tuple
 
     from pygls.lsp.client import BaseLanguageClient
 
@@ -51,7 +50,7 @@ async def rename(get_client_for):
     ],
 )
 async def test_prepare_rename(
-    rename: Tuple[BaseLanguageClient, types.InitializeResult],
+    rename: tuple[BaseLanguageClient, types.InitializeResult],
     path_for,
     uri_for,
     position: types.Position,
@@ -116,7 +115,7 @@ async def test_prepare_rename(
     ],
 )
 async def test_rename(
-    rename: Tuple[BaseLanguageClient, types.InitializeResult],
+    rename: tuple[BaseLanguageClient, types.InitializeResult],
     path_for,
     uri_for,
     position: types.Position,
